move built-in scalar functions
